Rename Renderful::Cache to Renderful::Cache::Base
